instrumentation: add new pass manager
The previous commit message was not correct. Indeed, it is possible to
to instrumentation with all kinds of passes. Only that the "new" pass
manager must be used (accessed by -fpass-plugin=).

The new build process is minimally invasive, since only one parameter
must be added to the compiler flags. The example can be executed with
the following

  make all-fast

Since the previous commit was not correct, the wrapper workaround is
removed in favor for this solution.

Signed-off-by: Eduard Vlad <eduard.vlad@rwth-aachen.de>